openai_dall-e/requirements.txt
openai
pika
Pillow
python-dotenv
requests